investment-ai-agent/requirements.txt
# Backend Dependencies
docling
scrapy
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
pandas
pydantic
requests
beautifulsoup4
watchdog
typing-extensions
